Tpetra: Added Details::congruent() to test if two comms are congruent.
"Congruent" here has the meaning of either MPI_IDENT and
MPI_CONGRUENT.  The former means that two MPI_Comm handles refer to
the same object.  The latter means that the two MPI_Comm have the same
number of processes in the same order.  I've extended this to its
logical meaning for a SerialComm (congruent to a Comm if and only if
the latter has one process).

This is useful for debugging.  I have debugging column Map union in
mind, but it's useful for other things as well.