Kokkos: Added "raw" sparse matrix add function (no GPU; CPU only).
The new function addSparseMatrices lives in the Kokkos::Raw namespace.
It computes the sum C = alpha*A + beta*B, where A and B are sparse
matrices in compressed sparse row format, and alpha and beta are
scalar coefficients.  C may not alias A or B.

This commit includes a (passing) test for different values of alpha
and beta.