Tpetra: CrsMatrix: insertGlobalValues no longer does column filtering.
insertGlobalValues used to "filter out" (silently ignore) columns not
in the column Map, if the matrix had a column Map.  This behavior made
it hard for users to debug their code, because it silently hid their
mistakes.  This commit changes this behavior, so that
insertGlobalValues inserts all the column indices given to it,
including those not in the column Map.  It also changes the test that
used to ensure that column Map filtering occurred.

This commit also adds a new private method insertGlobalValuesFiltered,
which retains the old column filtering behavior.  This is useful for
implementing DistObject methods, which currently require column
filtering in order to do redistribution correctly.  (That's because
the DistObject's Map is currently the matrix's row Map.)

Finally, this commit improves the public documentation of
insertGlobalValues and insertLocalValues.  It now gives performance
tips that recommend users to prefer local indices, give good upper
bounds on the number of entries in each row when constructing the
CrsMatrix, use StaticProfile if possible, and construct the CrsMatrix
with a const graph if the use case suggests it.

Note that I have not yet changed insertLocalValues, or any of
CrsGraph's insert*Indices methods, not to do column filtering.  That
will happen in later commits.