Kokkos Array: Fixed output of DeviceType name in performance test.
The KOKKOSARRAY_MACRO_TO_STRING(DeviceType) line in the performance
tests (which are templated on DeviceType) doesn't print the name of
the DeviceType template parameter, as it should.  Instead, it prints
"DeviceType" literally.  This is probably not what the authors
intended, so I changed the performance tests to take the name of
DeviceType as a string, and commented out the (now unused) macro.

Build/Test Cases Summary
Enabled Packages: KokkosArray
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=2,notpassed=0 (2.66 min)
3) SERIAL_RELEASE_SS => passed: passed=2,notpassed=0 (0.35 min)
Other local commits for this build/test group: 4a51100