Continue backwards compatibility work
Confirm that, not only do completed products build with -DSTK_MESH_DISALLOW_DEPRECATED_ENTITY_FNS,
but also confirm that all tests for those products pass when built with this flag. This exposed
a number of items that were missed.

Use templates to allow EntityGlobalIdComparator to take both stk::mesh::Entity and
Fmwk::MeshBulkData. The original name is now a typedef for the version that takes
stk::mesh::BulkData. This will allow clients to initialize containers that require
the comparator with Fmwk::MeshBulkDatas that haven't yet been attached to a stk::mesh::BulkData.
Clients should prefer the default if possible.

Aria was close to being free of backwards compatiblity use, so I finished it.

(From Sierra)