Panzer: Connecting DgDx to Epetra model evaluator, adding test
This was roughed in prior to this commit. However no test had been
implemented to excercise the capability. There were minor issues
within the ModelEvaluator that were causing problems (in particular
associated with the use of a pass by value method call, though in this
context pass by value is the right thing to do, methods were not beging
called in the right way). A simple reordering of method calls fixed
this issue.

Also, the test added excercises the mechanics of this capability but
does not check the values.