Ifpack2: Modified LocalFilter to allow alpha != 1, beta != 0 in the apply. Fixed BlockRelaxation::DoGaussSeidel.
Previously, LocalFilter's apply would only write A*x to y. Now the saxpy, or saxpby I guess,
is supported. Furthermore, BlockRelaxation doesn't work. Most specifically Gauss-Seidel.
I don't know about Jacobi yet. This is a fix of the DoGaussSeidel method and it works in
serial. There's a bug with AdditiveSchwarz. I'm working on it.