Ifpack2: Documented and reformatted BlockRelaxation.
Thanks to Tom Benson for implementing BlockRelaxation :-)  It's always
a bit scary to put a new class out there for users.  I decided to "lead
by doing" and illustrate documentation and code formatting standards for
Ifpack2, by using BlockRelaxation as an example and implementing those
standards.  Ifpack2 could use a lot of work in general, so I omitted the
more aggressive changes that would be more appropriate for a major
Ifpack2 refactor.

I also noticed that Ifpack2_BlockRelaxation_decl.hpp duplicated the
definition of the RelaxationType enum that was in
Ifpack2_Relaxation_decl.hpp.  To avoid duplicate definitions of this
enum, I moved it to Ifpack2_ConfigDefs.hpp.  I also put it in the
Ifpack2::Details namespace, since it is an implementation detail of
Ifpack2 and not meant for users.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2
0) MPI_DEBUG => passed: passed=123,notpassed=0 (3.65 min)
1) SERIAL_RELEASE => passed: passed=106,notpassed=0 (9.85 min)
Other local commits for this build/test group: 21da0ed