Alter stk-io to use a two-phase mesh-reading approach.
The first phase fills in mesh entities and connectivities but no
field-data.
The second phase fills in coordinates, attributes, etc.
The two-phase approach allows for improved efficiency, since the
first phase of filling in entities and connectivities doesn't have
to also copy field-data around when entities are moved to temporary
buckets etc.

This also allows for calling the create-edges code after the first
phase but before the second phase, which improves the performance of
create-edges.

nalu tests running on a blade show mostly small improvements due to this change.

Memory high-water-mark, mesh-read times, and edge-creation times all are improved,
but the improvements aren't very dramatic, at least for the nalu nightly tests
running on a blade.  Order of a few percent for the mesh-read and edge-creation timings,
and the high-water-mark numbers also show a few percent improvement for some tests
(but little or no improvement for others).

(From Sierra)