Ifpack2: Added OverlappingRowGraph class.
One issue with the current implementation of OverlappingRowMatrix is
that, even though it claims to implement Tpetra::RowMatrix, its
getGraph() method (that should return a Tpetra::RowGraph) is
unimplemented and throws an exception if called.  This motivated the
Ifpack2::Details::Tpetra_RowGraph class (that provides a graph "view"
of a Tpetra::RowMatrix) that I removed a few commits ago.  It's not
hard to write a Tpetra::RowGraph which works with the data in
OverlappingRowMatrix.  This commit adds Details::OverlappingRowGraph,
which implements Tpetra::RowGraph in this way.  I have not yet changed
OverlappingRowMatrix to create an instance of this class; that is the
logical next step.

OverlappingRowGraph is in the Details namespace for now, first because
I'm not certain about its interface (in particular, its constructor),
and second, because it is currently just an implementation detail of
OverlappingRowMatrix.

Note that OverlappingRowGraph is explicitly instantiated both for
GraphType=Tpetra::CrsGraph and for GraphType=Tpetra::RowGraph.
The CrsGraph explicit instantiation assumes the default Node and
LocalMatOps template parameters for Ifpack2's explicit instantiation
system.