KokkosArray: Fixed compile warnings.
The compile warnings referred to forward declarations using 'struct' or
'class', and the actual class declaration not matching ('class' vs.
'struct' for the forward declaration, or vice versa).  With Clang 3.2,
the warnings had the following form:

warning: class template 'ViewAssignment' was previously declared as a
struct template [-Wmismatched-tags]

This commit fixes those warnings.