Panzer: Updating driver header files to eventurally support ETI for the ME Factory
I tried to explicitly instantiate the ME Factory however that crashed
and burned because it was being built in the examples directory. However
it exposed the fact that many header files were missing from some of the
driver code. This fixes that issue. This is not an urgent commit but hopefully
will smooth the path to moving beyond the STK MEFactory Epetra code in the
examples directory and putting it in panzer_stk proper.