Stokhos:  Add COO-inspired Kokkos SG tensor and multiply kernel.
This uses a COO-like pattern of paralllelizing over sparse Cijk nonzeros
(as opposed to rows like the CRS pattern), and uses a custom segmented
reduction within shared memory to compute the update for each row.
This provides better performance for smaller problems with fewer
nonzeros per row compared to the CRS approach, but has lower peak
performance.  It also suffers from the same occupancy issues as the CRS
kernel, but I am hopeful it will be more amenable to tiling of the Cijk
tensor.