Claps: Fix issue with std namespace
The change to Epetra removing the using namespace std from a header file caused
a number of claps files to fail to build because of unqualified endl and cout
statements. Since claps is only experimental this wasn't caught. However Sierra
uses claps so this showed itself during the integration process.