MueLu: BrickAggregationF: should now work (but only with 3x3)
The current implementation cannot deal with bricks larger than 3x3(x3).
The reason is that aggregation infrastructure in place has major drawbacks.

Aggregates class is constructed with a help of a provided map, either taken from
a graph, or provided directly. This map is usually taken to be a column map of a
matrix. The reason for that is that if we have an overlapped aggregation, we want
the processor owning aggregates to store agg id for all nodes in this aggregate.
If we used row map, there would be no way for the processor to know whether there
are some other nodes on a different processor which belong to its aggregate.
On the other hand, using column map allows both vertex2AggId and procWinner arrays
in Aggregates class to store some extra data, such as whether nodes belonging to a
different processor belong to this processor aggregate.
The drawback of this is that it stores only overlap=1 data. For aggressive coarsening,
such a brick aggregation with a large single dimension of brick, it could happen
that we need to know depth two or more extra nodes in the other processor subdomain.

Another issue is that we may have some implicit connection between aggregate map and
maps of A used in the construction of a tentative prolongator.

Another issue is that it seems that some info is unused or not required. Specifically,
it seems that if a node belongs to an aggregate on a different processor, we don't
actually need to set vertex2AggId and procWinner, despite the following comment in
Aggregates decl:
     vertex2AggId[k] gives a local id
     corresponding to the aggregate to which
     local id k has been assigned.  While k
     is the local id on my processor (MyPID)
     vertex2AggId[k] is the local id on the
     processor which actually owns the
     aggregate. This owning processor has id
     given by procWinner[k].
It is possible that that info is only used during arbitration in CoupledAggregationFactory.

The steps that we need to do to resolve this issue:
  - Break the link between maps in TentativePFactory, allowing any maps in Aggregates
  - Allow Aggregates to construct their own maps, if necessary, OR
  - construct aggregates based on row map